requirements.txt
pytest>=8.2.0
deepeval>=1.0.0
python-dotenv>=1.0.1